+++ src/Identity/Lazy.php
@@ -3,7 +3,10 @@ declare(strict_types = 1);
 
 namespace Innmind\Immutable\Identity;
 
-use Innmind\Immutable\Identity;
+use Innmind\Immutable\{
+    Identity,
+    Sequence,
+};
 
 /**
  * @psalm-immutable
@@ -37,6 +40,11 @@ final class Lazy implements Implementation
 
         /** @psalm-suppress ImpureFunctionCall */
         return Identity::lazy(static fn() => $map($value())->unwrap());
+    }
+
+    public function toSequence(): Sequence
+    {
+        return Sequence::lazy(fn() => yield $this->unwrap());
     }
 
     public function unwrap(): mixed